баги :navbar не обновляется после авторизации, myselect

+++ frontend/frontend/src/components/TodoItems.jsx
@@ -5,6 +5,8 @@ import MyModal from './UI/MyModal/MyModal';
 import PostForm from './PostForm';
 const ToDoItem = ({post,value}) => {
     const [modalupdate, setModalUpdate] = useState(false);
+    const token = localStorage.getItem('token');
+
     return (
         <div className = "items">
             <strong>{post.title}</strong>
@@ -20,7 +22,7 @@ const ToDoItem = ({post,value}) => {
                 <PostForm value={value} option={value} defaultValue='Категория' flag={true } idPutToDO={post.id} />
             </MyModal>
             <div>
-                <button type="submit" onClick={() => { TodoService.deleteTodo(post.id); document.location.reload(); }}>Удалить</button>
+                <button type="submit" onClick={() => { TodoService.deleteTodo(post.id,{Authorization: 'Token '+ JSON.parse(token) }); document.location.reload(); }}>Удалить</button>
             </div>
         </div>
     )
